requirements.txt
python-telegram-bot>=21.0
openai>=1.0
python-dotenv
apscheduler>=3.10
